backend/requirements.txt
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.1
flask-cors==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
MetaTrader5==5.0.5120
numpy==2.3.1
openpyxl==3.1.5
pandas==2.3.1
python-dateutil==2.9.0.post0
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
twelvedata==1.2.25
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
